trip_snatchers_backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-dateutil==2.8.2
APScheduler==3.10.4
python-dotenv==1.0.1 